+++ Sourcery/Generating/Template/Stencil/StencilTemplate.swift
@@ -28,6 +28,7 @@ final class StencilTemplate: StencilSwiftKit.StencilSwiftTemplate, Template {
     private static func sourceryEnvironment() -> Stencil.Environment {
         let ext = Stencil.Extension()
         ext.registerFilter("upperFirst", filter: Filter<String>.make({ $0.upperFirst() }))
+        ext.registerFilter("lowerFirst", filter: Filter<String>.make({ $0.lowerFirst() }))
         ext.registerFilterWithTwoArguments("replace", filter: { (source: String, substring: String, replacement: String) -> Any? in
             return source.replacingOccurrences(of: substring, with: replacement)
         })
@@ -167,6 +168,11 @@ extension String {
         return first + other
     }
 
+    fileprivate func lowerFirst() -> String {
+        let first = String(characters.prefix(1)).lowercased()
+        let other = dropFirst()
+        return first + other
+    }
 }
 
 private struct Filter<T> {