+++ Sourcery/Utils/Path+Extensions.swift
@@ -33,11 +33,19 @@ extension Path {
     var isTemplateFile: Bool {
         return self.extension == "stencil" ||
             self.extension == "swifttemplate" ||
-            self.extension == "js"
+            self.extension == "ejs"
     }
 
     var isSwiftSourceFile: Bool {
         return !self.isDirectory && self.extension == "swift"
+    }
+
+    init(_ string: String, relativeTo relativePath: Path) {
+        var path = Path(string)
+        if !path.isAbsolute {
+            path = (relativePath + path).absolute()
+        }
+        self.init(path.string)
     }
 
 }