+++ SourceryTests/Generating/JavaScriptTemplateSpecs.swift
@@ -12,10 +12,30 @@ class JavaScriptTemplateTests: QuickSpec {
                 return Stubs.cleanTemporarySourceryDir()
             }()
 
-            let templatePath = Stubs.jsTemplates + Path("Equality.js")
-            let expectedResult = try? (Stubs.resultDirectory + Path("Basic.swift")).read(.utf8)
-
             it("generates correct output") {
+                let templatePath = Stubs.jsTemplates + Path("Equality.ejs")
+                let expectedResult = try? (Stubs.resultDirectory + Path("Basic.swift")).read(.utf8)
+
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources([Stubs.sourceDirectory]), usingTemplates: [templatePath], output: outputDir) }.toNot(throwError())
+
+                let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
+                expect(result).to(equal(expectedResult))
+            }
+
+            it("handles includes") {
+                let templatePath = Stubs.jsTemplates + Path("Includes.ejs")
+                let expectedResult = try? (Stubs.resultDirectory + Path("Basic+Other.swift")).read(.utf8)
+
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources([Stubs.sourceDirectory]), usingTemplates: [templatePath], output: outputDir) }.toNot(throwError())
+
+                let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
+                expect(result).to(equal(expectedResult))
+            }
+
+            it("handles includes from included files relatively") {
+                let templatePath = Stubs.jsTemplates + Path("SubfolderIncludes.ejs")
+                let expectedResult = try? (Stubs.resultDirectory + Path("Basic.swift")).read(.utf8)
+
                 expect { try Sourcery(cacheDisabled: true).processFiles(.sources([Stubs.sourceDirectory]), usingTemplates: [templatePath], output: outputDir) }.toNot(throwError())
 
                 let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))